configury: support non-autotools LuaRocks installation.
* build-aux/mkrockspecs.in: Add a Bourne shell preamble to make
a PATH search for suitable Lua binary and re-execute.
Support new --module-dir command-line option to specify a tree
of `.lua` files for installation, in preference to a full auto-
tools build.
Collect and correctly quote a LuaRocks `build.modules` table in
the generated rockspec when --module-dir is passed.
Also, await rockspec configuration on standard input when new
positional parameter for FILE is '-'.
* specs/package_spec.yaml: Remove dummy spec file.
* specs/mkrockspecs_spec.yaml: Verify behaviour of new functions
in mkrockspecs.
* specs/specs.mk (specl_SPECS): Adjust.
* build-aux/mkrockspecs.mk ($(package_rockspec)): Support optional
mkrockspecs_args for user arguments to mkrockspecs script.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>