bootstrap: allow pre-emptive APPVAR to skip rockspec version checks.
Avoid a catch-22 where you can't run bootstrap because
prerequisite rocks are missing, but you can't install the rocks
until you have the configured Makefiles necessary to generate
luarocks-config.lua.
* build-aux/bootstrap.slingshot (slingshot_check_rockspecs): If
APPVARs (e.g. LDOC, SPECL) are already set in the environment,
but the paths they point to don't yet exist, skip the version
check, but show a warning that configure and/or make will be
needing them to work properly.
* bootstrap: Regenerate.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>