bootstrap: always run `git submodule update -- slingshot`.
If we are in a git directory, and there is a git executable in
the user's PATH, then always synchronize the slingshot submodule
in case we changed revisions since the last bootstrap run.
* build-aux/bootstrap.slingshot
(slingshot_require_slingshot_submodule): Move the $GIT calls and
dependencies out of the inner if, so that it always triggers
when we have a git binary.
* bootstrap: Regenerate.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>