bootstrap: don't run `git submodule update` outside working dir.
5cd707a was not quite correct, and would sometimes try to update
submodules in a tarball source tree.
* build-aux/bootstrap.slingshot
(slingshot_require_slingshot_submodule): If there is a .gitmodules
file, make sure it contains the slingshot submodule, and update
it to the correct version for the parent revision; incidentally
only doing a full clone of the submodule along the way if it is
entirely absent.
* bootstrap: Regenerate.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>