bootstrap: detect out of date bootstrap with slingshot clients.
* build-aux/bootstrap.slingshot (require_bootstrap_uptodate): A
third version of this function!  The main one from upstream
bootstrap is overridden by ./bootstrap.conf to merge in the extra
functions in this file.  The new one here is to automate checking
for out of date bootstrap in slingshot client projects when they
update their slingshot submodule and rebootstrap.
* bootstrap: Regenerate.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>