travis: force install git rockspecs over older dependency rocks.
This works around a bug where luarocks pulls the most recent
stdlib rock from the repo as a dependency of specl, but then
refuses to remove it when installing latest unreleased stdlib
in preparation for installcheck.  In that case, the spec files
from the unreleased stdlib are used to check the last release...
which will almost certainly fail, since the unreleased stdlib
will have tests for new and unreleased APIs!
* travis.yml.in (install): In order for installcheck to work
without a ton of hoop-jumping and special-casing, assume that we
want backwards compatibility, and force install the git rockspec
over any older version installed as a dependency rock earlier on.
* .travis.yml: Regenerate.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>